requirements.txt
flask==3.0.0
flask-cors==4.0.0
supabase==2.3.0
yfinance==0.2.36
requests==2.31.0
APScheduler==3.10.4
gunicorn==21.2.0
beautifulsoup4==4.12.3
schedule==1.2.1
python-dotenv==1.0.0
pytz==2024.1
lxml==5.1.0
alpaca-trade-api==3.2.0
redis==5.0.1
anthropic==0.25.0
google-auth==2.29.0
scipy>=1.12.0
numpy>=1.26.0
pandas>=2.2.0
pyyaml>=6.0
py_vollib>=1.0.1
